GP131: Update TEE_AsymmetricEncrypt() and TEE_AsymmetricDecrypt()
Updates TEE_AsymmetricEncrypt() and TEE_AsymmetricDecrypt() according to
TEE Internal Core API version 1.3.1. The srcLen and dstLen parameters
are changed to use size_t instead of uint32_t

Adds the optional attribute TEE_ATTR_RSA_OAEP_MGF_HASH which if provided
specifies the MGF1 hash function to use. Prior to GP v1.3.1 it was
implementation defined. OP-TEE only supports using the same MGF1 hash
function as the internal hash function of the algorithim, that is, for
the algorithm TEE_ALG_RSAES_PKCS1_OAEP_MGF1_x that hash function
TEE_ALG_x is used.

+++ lib/libutee/include/tee_internal_api.h
@@ -423,8 +423,8 @@ TEE_Result __GP11_TEE_AEDecryptFinal(TEE_OperationHandle operation,
 TEE_Result TEE_AsymmetricEncrypt(TEE_OperationHandle operation,
 				 const TEE_Attribute *params,
 				 uint32_t paramCount, const void *srcData,
-				 uint32_t srcLen, void *destData,
-				 uint32_t *destLen);
+				 size_t srcLen, void *destData,
+				 size_t *destLen);
 TEE_Result __GP11_TEE_AsymmetricEncrypt(TEE_OperationHandle operation,
 					const __GP11_TEE_Attribute *params,
 					uint32_t paramCount,
@@ -434,8 +434,8 @@ TEE_Result __GP11_TEE_AsymmetricEncrypt(TEE_OperationHandle operation,
 TEE_Result TEE_AsymmetricDecrypt(TEE_OperationHandle operation,
 				 const TEE_Attribute *params,
 				 uint32_t paramCount, const void *srcData,
-				 uint32_t srcLen, void *destData,
-				 uint32_t *destLen);
+				 size_t srcLen, void *destData,
+				 size_t *destLen);
 TEE_Result __GP11_TEE_AsymmetricDecrypt(TEE_OperationHandle operation,
 					const __GP11_TEE_Attribute *params,
 					uint32_t paramCount,
